requirements.txt
torch==2.4.1
torchvision==0.19.1
tensorboard==2.14.0
biopython==1.83
tqdm==4.67.1
scikit-learn==1.3.2
matplotlib==3.7.5
seaborn==0.13.2
jupyter==1.1.1
